+++ retinify/src/stream.hpp
@@ -4,7 +4,7 @@
 #pragma once
 
 #include "retinify/attributes.hpp"
-#include "retinify/constraints.hpp"
+#include "retinify/nocopymove.hpp"
 #include "retinify/status.hpp"
 
 #ifdef BUILD_WITH_TENSORRT
